% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-plot.R
\name{plot_to_img_content}
\alias{plot_to_img_content}
\alias{plot_to_img_content.plotly}
\alias{plot_to_img_content.ggplot}
\title{Convert a plot object to a PNG image content}
\usage{
plot_to_img_content(p)

\method{plot_to_img_content}{plotly}(p)

\method{plot_to_img_content}{ggplot}(p)
}
\arguments{
\item{p}{The plot object; currently, plotly and ggplot2 are supported.}
}
\value{
An ellmer content_image object
}
\description{
Convert a plot object to a PNG image content
}
\keyword{internal}
